IA-generativa/generacion_multimodal_contenido/generacion_multimodal_contenido/requirements.txt
# requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.18.0
Pillow>=9.0.0
numpy>=1.24.0
soundfile>=0.12.0

# API y web
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
streamlit>=1.25.0
requests>=2.31.0
pydantic>=2.0.0

# Utilidades
scipy>=1.10.0
accelerate>=0.20.0

#xformers>=0.0.20 